examples/enterprise-demo/pyproject.toml
[project]
name = "enterprise-demo"
version = "0.1.0"
description = "企业级日志示例 - 展示 yai-loguru-sinks 的 SLS 集成功能"
authors = [
    {name = "YAI Team", email = "team@yai.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "loguru>=0.7.0",
    "yai-loguru-sinks",
    "loguru-config>=0.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[tool.uv.sources]
yai-loguru-sinks = { workspace = true }